+++ CHANGELOG.md
@@ -12,6 +12,8 @@
 - Bumped vitest and playwright packages
 - Fixed: GetClientIP now handles comma-separated X-Forwarded-For headers and updated reverse proxy documentation
 - Added last login information to the users list and aligned items in users and projcts list
+- Added an admin password legitimation step for a project deletion attempt
+- Added an admin password legitimation step when an admin wants to activate or deactive a user
 
 # 1.3.1